+++ src/chunker.js
